+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+    "editor.tabSize": 4,
+    "prettier.printWidth": 120
+
+}
