+++ src/main/java/duke/dukeexception/DukeException.java
@@ -1,3 +1,5 @@
+package dukeexception;
+
 public class DukeException extends Exception{
     public DukeException(String message) {
         super(message);
