+++ src/main/java/duke/dukeexception/EmptyTaskArgumentException.java
@@ -1,3 +1,5 @@
+package dukeexception;
+
 public class EmptyTaskArgumentException extends DukeException {
     public EmptyTaskArgumentException(String message) {
         super(message);
