+++ src/main/java/duke/dukeexception/MarkIndexDoesNotExistException.java
@@ -1,3 +1,5 @@
+package dukeexception;
+
 public class MarkIndexDoesNotExistException extends DukeException {
     public MarkIndexDoesNotExistException(String message) {
         super(message);
