+++ src/main/java/duke/dukeexception/UnmarkIndexDoesNotExistException.java
@@ -1,3 +1,5 @@
+package dukeexception;
+
 public class UnmarkIndexDoesNotExistException extends DukeException {
     public UnmarkIndexDoesNotExistException(String message) {
         super(message);
