+++ src/main/java/duke/task/Deadline.java
@@ -1,5 +1,7 @@
+package task;
+
 public class Deadline extends Task {
-    public String by;
+    private String by;
 
     public Deadline(String description, String by) {
         super(description);
